training/requirements.txt
torch>=2.4.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
yfinance>=0.2.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
onnx>=1.15.0
onnxruntime>=1.16.0
ta>=0.11.0
optuna>=3.0.0
flask>=3.0.0  # For Hyperparameter Tuning (The "Surgical" edge)
